Major code moves.
Making bridge_converter much more manageable by abstracting out
the hard bits into foreign_mod_converter.
